mcp-servers/image-generation/server-simple.py
#!/usr/bin/env python3
"""
Simplified MCP Server for Image Generation
Shows installation status and instructions if dependencies missing
"""

import asyncio
import sys
from mcp.server import Server
from mcp.types import Tool, TextContent

# Check dependencies
try:
    import torch
    from diffusers import DiffusionPipeline
    from PIL import Image
    HAS_DEPS = True
    DEP_STATUS = "All dependencies installed"
except ImportError as e:
    HAS_DEPS = False
    DEP_STATUS = f"Missing dependencies: {e}"
    print(f"WARNING: {DEP_STATUS}", file=sys.stderr)
    print("Run: pip install torch torchvision diffusers transformers accelerate Pillow", file=sys.stderr)

app = Server("image-generation-simple")

@app.list_tools()
async def list_tools() -> list[Tool]:
    return [
        Tool(
            name="generate_image_free",
            description="Generate images locally using SDXL (FREE - no API keys). "
                       "Returns installation instructions if dependencies not installed.",
            inputSchema={
                "type": "object",
                "required": ["prompt"],
                "properties": {
                    "prompt": {
                        "type": "string", 
                        "description": "Image description"
                    },
                    "model": {
                        "type": "string",
                        "enum": ["sdxl-lightning", "sdxl-turbo"],
                        "default": "sdxl-lightning"
                    }
                }
            }
        ),
        Tool(
            name="check_image_gen_status",
            description="Check if image generation dependencies are installed",
            inputSchema={"type": "object", "properties": {}}
        )
    ]

@app.call_tool()
async def call_tool(name: str, arguments: dict) -> list:
    if name == "check_image_gen_status":
        if HAS_DEPS:
            return [TextContent(type="text", text="✅ Image generation ready!\n\nDependencies installed:\n- PyTorch: " + torch.__version__ + "\n- CUDA: " + str(torch.cuda.is_available()) + "\n- Diffusers: Installed\n\nYou can now generate images!")]
        else:
            return [TextContent(type="text", text="⚠️ Image generation NOT ready\n\nMissing dependencies. Install with:\n\npip install torch torchvision --index-url https://download.pytorch.org/whl/cu121\npip install diffusers transformers accelerate Pillow\n\nRequires NVIDIA GPU with 8GB+ VRAM.")]
    
    if name == "generate_image_free":
        if not HAS_DEPS:
            return [TextContent(type="text", text="⚠️ Cannot generate image - dependencies not installed\n\nTo enable FREE image generation, run:\n\nC:\\Users\\Tyson\\.openclaw\\install-all-deps.bat\n\nOr manually:\npip install torch torchvision --index-url https://download.pytorch.org/whl/cu121\npip install diffusers transformers accelerate Pillow\n\nRequirements:\n- NVIDIA GPU with 8GB+ VRAM\n- CUDA 12.1+\n- ~10GB disk space for models")]
        
        prompt = arguments.get("prompt", "")
        model = arguments.get("model", "sdxl-lightning")
        
        return [TextContent(type="text", text=f"✅ Dependencies installed!\n\nTo generate: '{prompt}'\n\nModel: {model}\n\nFirst generation will download ~5GB model files.\nThis may take 5-10 minutes on first use.")]
    
    raise ValueError(f"Unknown tool: {name}")

async def main():
    from mcp.server.stdio import stdio_server
    print("Image Generation MCP Server (Simple)", file=sys.stderr)
    print(f"Status: {DEP_STATUS}", file=sys.stderr)
    
    async with stdio_server() as (read_stream, write_stream):
        await app.run(read_stream, write_stream, app.create_initialization_options())

if __name__ == "__main__":
    asyncio.run(main())
